+++ R/multipleGroup.R
@@ -375,6 +375,7 @@
 #' mod_configural <- multipleGroup(dat, model, group = group)
 #' plot(mod_configural)
 #' plot(mod_configural, type = 'SE')
+#' plot(mod_configural, type = 'gen.difficulty')
 #' itemplot(mod_configural, 1)
 #' itemplot(mod_configural, 1, type = 'info')
 #' plot(mod_configural, type = 'trace') # messy, score function typically better
@@ -437,6 +438,7 @@
 #' summary(mod_mix)
 #' plot(mod_mix)
 #' plot(mod_mix, type = 'trace')
+#' plot(mod_mix, type = 'gen.difficulty')
 #' itemplot(mod_mix, 1, type = 'info')
 #'
 #' head(fscores(mod_mix)) # theta estimates